ISPN-10396 Refactor: create one test per resource
