+++ server/rest/src/test/java/org/infinispan/rest/resources/CacheResourceOffHeapTest.java
@@ -1,11 +1,11 @@
-package org.infinispan.rest;
+package org.infinispan.rest.resources;
 
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.configuration.cache.StorageType;
 import org.testng.annotations.Test;
 
-@Test(groups = "functional", testName = "rest.RestOperationsOffHeapTest")
-public class RestOperationsOffHeapTest extends BaseRestOperationsTest {
+@Test(groups = "functional", testName = "rest.CacheResourceOffHeapTest")
+public class CacheResourceOffHeapTest extends BaseCacheResourceTest {
 
    @Override
    public ConfigurationBuilder getDefaultCacheBuilder() {
